feat: Overlay Refactor Part 2: Port all compatible overlays to TextOverlay (#1278)
* feat: Overlay Refactor Part 1: Introduce TextOverlay

* fix: Refactor to be actually useful

* fix: Fix up i18n

* fix: Move all rendering to TextOverlay

* feat: Port ArrowShieldTrackingFeature to TextOverlay

* feat: Port AuraTimerOverlay to TextOverlay

* feat: Port ShamanMaskOverlay to TextOverlay

* feat: Port ShamanTotemTimerOverlay to TextOverlay

* feat: Port MobTotemTimerOverlay to TextOverlay

* fix: Fix unnecessary long function name

* fix: Use new function options, fixes

* fix: Review fixes

* fix: Fix merge and sort

* fix: remove comment

* ci: spotless formatting

---------

+++ common/src/main/java/com/wynntils/models/mobtotem/MobTotem.java
@@ -35,9 +35,9 @@ public class MobTotem {
         this.timerString = timerString;
     }
 
-    public int getDistanceToPlayer() {
+    public double getDistanceToPlayer() {
         // y() - 4 because the entity is 4 blocks above the ground
-        return (int) Math.sqrt(McUtils.player().distanceToSqr(new Vec3(location.x(), location.y() - 4, location.z())));
+        return Math.sqrt(McUtils.player().distanceToSqr(new Vec3(location.x(), location.y() - 4, location.z())));
     }
 
     public double getLookAngleDiff() {
